Minor refactoring and fixing tests
